Rough cut of license zones
Each license is put in its own zone area and stacked diagonally. They
all sit in a flex container and are hidden by default until a license is
added. Stacking looks okay but it's all a bit clumsy. The idea is there,
but CSS needs some clean up and JS objects need to be positioned on the
aution when created.